refactor : 사용되지 않은 파일 제거
